Collision Update
2 different collision handling attempts here, both work relatively well
but can still result in being able to move through the wall in some
weird situations. Both work the same way, by causing the player to move
backwards on a collision, but break differently. The key to making -8.0
speed work was switching the wall to a sensor physics object. Sensors
detect collision but leave the handling of the collision to the
programmer. Still some clean up to be done here.