Sprite Added w/Motion
Problems:
-Only 4 directions of motion currently (and no clear attack motion)
-Currently only playing 2 of the 3 frames it should play for each
direction. Might be a timing problem.

If you guys find any cool sprites that aren't knock offs of a famous
game(most o the ones I saw online were) send them to me, implementing a
different sprite in place of this one would not be hard at all. Just
getting the ball rolling on the sprite front and familiarizing myself
with the interface.